add docstrings and rename struck_scaler
